Homework 4.2 sql files is created

+++ src/main/java/ru/hogwarts/school/controller/StudentController.java
@@ -79,7 +79,7 @@ public class StudentController {
     }
 
     @GetMapping("/info/count")
-    public Integer getCountStudents() {
+    public Long getCountStudents() {
         return studentService.countOfStudents();
     }
